remove i32

+++ crates/jpreprocess-core/src/accent_rule.rs
@@ -96,11 +96,11 @@ impl Display for AccentType {
 #[derive(Clone, PartialEq, Serialize, Deserialize, Debug)]
 pub struct ChainRule {
     pub accent_type: AccentType,
-    pub add_type: i32,
+    pub add_type: isize,
 }
 
 impl ChainRule {
-    pub fn new(accent_type: AccentType, add_type: i32) -> Self {
+    pub fn new(accent_type: AccentType, add_type: isize) -> Self {
         Self {
             accent_type,
             add_type,
